chore: update based on feedback (move escape to util, add test)
