+++ src/util.c
@@ -1621,3 +1621,25 @@ void writePointerWithPadding(unsigned char *buf, const void *ptr) {
     /* if it is 32-bit system, pad the remaining 4 bytes with zero */
     if (ptr_size == 4) memset(buf + ptr_size, 0, ptr_size);
 }
+
+/*
+ * Escape a Unicode string for JSON output, following RFC 7159:
+ * https://datatracker.ietf.org/doc/html/rfc7159#section-7
+ */
+sds escapeJsonString(sds s, const char *p, size_t len) {
+    s = sdscatlen(s, "\"", 1);
+    while (len--) {
+        switch (*p) {
+        case '\\':
+        case '"': s = sdscatprintf(s, "\\%c", *p); break;
+        case '\n': s = sdscatlen(s, "\\n", 2); break;
+        case '\f': s = sdscatlen(s, "\\f", 2); break;
+        case '\r': s = sdscatlen(s, "\\r", 2); break;
+        case '\t': s = sdscatlen(s, "\\t", 2); break;
+        case '\b': s = sdscatlen(s, "\\b", 2); break;
+        default: s = sdscatprintf(s, *(unsigned char *)p <= 0x1f ? "\\u%04x" : "%c", *p);
+        }
+        p++;
+    }
+    return sdscatlen(s, "\"", 1);
+}
